Normalize, csv, joblib e onnx mais atualizados
